+++ src/lib/waku_light_push/index.ts
@@ -53,7 +53,7 @@ export class WakuLightPush {
 
   async push(
     encoder: Encoder,
-    message: Message,
+    message: Partial<Message>,
     opts?: PushOptions
   ): Promise<SendResult> {
     const pubSubTopic = opts?.pubSubTopic ? opts.pubSubTopic : this.pubSubTopic;
@@ -79,7 +79,7 @@ export class WakuLightPush {
     const recipients: PeerId[] = [];
 
     try {
-      const protoMessage = await encoder.encodeProto(message);
+      const protoMessage = await encoder.toProtoObj(message);
       if (!protoMessage) {
         log("Failed to encode to protoMessage, aborting push");
         return { recipients };